Create Baek Joon 2869
To reach the top, there are two ways.
1. Perfectly reach the point. (ex. V=1, A=1, B=0)
2. Easily reach the top. (ex. V=1, A=2, B=0)
Let's say the snail will move A and slip B for N days and reach V at N+1 day.  Then
V <=(N+1)A-N*B
(V-A)/(A-B) <= N
In the way of 1. (V-A)/(A-B) days will be spent, and in the way of 2. (V-A)/(A-B)+1 days will be spent.
Also, the day should be counted form day 1, so 1 will be added to the result.